Sanitise recipient addresses before sending
In case an address is added in the form of `<someone@example.com>`, the `<>` will be removed. If It is `Someone <someone@example.com>`, It will be parsed by extracting the same and assigning it appropriately. In case of `recipients=[{"address": "Someone <someone@example.com>", "name": "Some other name"}],`, `Some other name` will be prioritised.